==> code/3d_nDEGs.Rmd <==
---
title: "Number of DE Genes"
author: "Zhaoheng Li, Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r,warning = F, message= F}
 library(Seurat)
library(ggpubr)
library(viridis)
library(tidyr)
library(dplyr)
library(tidyr)
library(readxl)
library(stringr)
library(gridExtra)
library(grid)
library(purrr)
library(tidyverse)
out_dir = "/fh/fast/sun_w/zhaoheng_li/MSK_DEC24/Analysis/outs/differential_expression"
```


# save DESeq2 results to tsv files

```{r}

for (tp in c("D-1", "D3", "D7", "D11", "D18")) {
  res <- readRDS(file.path(out_dir, paste0("DE_pseudobulk_", tp, ".RDS")))
  write.table(
    res,
    file      = file.path(out_dir, paste0("DE_pseudobulk_", tp, ".tsv")),
    sep       = "\t",
    quote     = FALSE,
    row.names = FALSE,
    col.names = TRUE
  )
  
}

```


# barplots

```{r barplot,fig.width=8,fig.height=6}

for( tp in c("D-1","D3","D7","D11","D18")){
  print(tp)
  res = readRDS(file.path(out_dir,paste0( "DE_pseudobulk_",tp,".RDS")))
   
  fig = res%>%
   dplyr::filter(padj<0.05)%>%
      group_by(celltype,genotype)%>%
   dplyr::summarise(n_DEG = n_distinct(gene))%>%
    ggplot(aes(x=genotype,y=n_DEG))+
    geom_col()+
    xlab("")+
    ylab("Number of DE Genes")+
    theme_classic()+
    theme(text = element_text(size = 10),
            axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1))+
    facet_wrap(~celltype)
  ggsave(file.path(out_dir,paste0("nDEGs/",tp,"_nDEG.png",sep='')),fig,width = 8,height = 6)
  print(fig)
}
```

# add sample sizes

```{r}
f_pseudobulk_list = file.path(out_dir,"pseudobulk_list.RDS")
pseudo_obj_list  = readRDS(f_pseudobulk_list)

```

```{r}
min_cells = 30


pseudo_obj_list = lapply(pseudo_obj_list,function(pseudo_obj){
  pseudo_obj%>%subset(subset = n>=min_cells)
})
```


```{r,fig.width=8,fig.height=6}

nDEG = list()
for( tp in c("D-1","D3","D7","D11","D18")){
  print(tp)
  res = readRDS(file.path(out_dir,paste0( "DE_pseudobulk_",tp,".RDS"))) 
  df = res%>%
   dplyr::filter(padj<0.05)%>%
      group_by(celltype,genotype)%>%
   dplyr::summarise(n_DEG = n_distinct(gene))
  df$tp = tp
  nDEG[[tp]] = df
}
nDEG_df = Reduce(rbind, nDEG)

nDEG_df
```


```{r}

meta_counts = imap_dfr(
  pseudo_obj_list,
  ~{
    tp   = .y                
    meta = .x[[]]             
                              
    meta %>%
      group_by(celltype, genotype) %>%       
      summarise(
        n_pseudobulk = n(),                     
        n_cells_tol   = sum(n),                
        .groups   = "drop"
      ) %>%
      mutate(tp = tp)                       
  }
)


wt_counts = meta_counts %>%
  filter(genotype == "WT") %>%
  select(tp, celltype, 
         n_WT_samples  = n_pseudobulk,
         n_cells_WT    = n_cells_tol)
ko_counts = meta_counts %>%
  filter(genotype != "WT") %>%   # every non-WT KO
  select(tp, celltype, genotype, 
         n_KO_samples = n_pseudobulk,
         n_cells_KO   = n_cells_tol)


nDEG_df = nDEG_df %>%                
  left_join(wt_counts, by = c("tp", "celltype")) %>%         
  left_join(ko_counts, by = c("tp", "celltype", "genotype"))  
```

```{r}
nDEG_df = nDEG_df %>% 
  mutate(
    id        = paste(tp, celltype, sep = "_"),
    log_nDEG  = log10(n_DEG + 1)
  )

nDEG_df$id = fct_inorder(nDEG_df$id)
nDEG_df
```



# dotplot

```{r,fig.height=6,fig.width=6}
id_levels = levels(nDEG_df$id)      
df = nDEG_df %>%
  filter(tp != "D-1") %>%                    
  mutate(id = factor(id, levels = id_levels)) %>% 
  droplevels()                                    

dotplot = df %>% ggplot(aes(x = id, y = genotype,
           size  = log_nDEG,
           colour = log_nDEG)) +
  geom_point(alpha = 0.85) +
  scale_size(range = c(0.8, 5), name = "log10(#DEG + 1)") +
  scale_colour_gradient(
    low  = "white",
    high = "red",
    name = "log10(#DEG + 1)"
  )+
  scale_x_discrete(limits = levels(df$id)) +
  labs(x = NULL, y = "KO genotype") +
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1),
    legend.position = "right"
  )
dotplot
ggsave(file.path(out_dir,"nDEGs/nDEGs_dotplot.png"),dotplot,width = 6,height = 6)
```

```{r}

write.table(
  nDEG_df,
  file      = file.path(out_dir,"nDEGs/nDEG.tsv"),
  sep       = "\t",
  quote     = FALSE,
  row.names = FALSE
)
```


# Session information

```{r}
gc()
sessionInfo()
```
