==> code/3g_compare-DE.Rmd <==
---
title: "Compare Pseudobulk and Cell Level DE Results"
author: "Zhaoheng Li, Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r,warning = F, message= F}
 library(Seurat)
library(ggpubr)
library(viridis)
library(tidyr)
library(dplyr)
library(tidyr)
library(readxl)
library(stringr)
library(gridExtra)
library(grid)
library(purrr)
library(tidyverse)
out_dir = "/fh/fast/sun_w/zhaoheng_li/MSK_DEC24/Analysis/outs/differential_expression"
celllevel_out_dir = file.path(out_dir,"cell-level")
pseudobulk_out_dir = out_dir
fig_dir =  file.path(out_dir,"compare-DE")
```



```{r,fig.height=12,fig.width=12}
for(tp in c("D-1","D3","D7","D11","D18")){
  print(paste0("time point: ", tp))
  pseudobulk_res = read.delim(file.path(pseudobulk_out_dir, paste0("DE_pseudobulk_", tp, ".tsv")), 
                   header = TRUE, sep = "\t", stringsAsFactors = FALSE)
  
  wilcox_res = read.csv(file.path(celllevel_out_dir,paste0("DE_cell-level_",tp,".csv")),row.names = 1)
  merged <- pseudobulk_res %>%
    select(celltype, genotype, gene, pvalue) %>%
    inner_join(
      wilcox_res %>% select(celltype, genotype, gene, p_val),
      by = c("celltype", "genotype", "gene")
    ) %>%
    mutate(
      pvalue   = pmax(pvalue, .Machine$double.xmin),
      p_val    = pmax(p_val,  .Machine$double.xmin),
      nl10_pb  = -log10(pvalue),
      nl10_wc  = -log10(p_val)
    )
  for(ct in unique(merged$celltype)){
    print(paste0("cell type: ", ct))
    fig = merged %>%
      filter(celltype == ct) %>%
      ggplot(aes(x = nl10_pb, y = nl10_wc)) +
      geom_point(alpha = 0.6, size = 1) +
      geom_abline(slope = 1, intercept = 0, linetype = "dashed") +
      stat_cor(method = "spearman", label.x = 1, label.y = 19, size = 3) + 
      coord_cartesian(xlim = c(0, 20), ylim = c(0, 20)) +
      labs(
        x = "-log10(pseudobulk p-val)",
        y = "-log10(wilcoxon p-val)",
        title = paste(tp, ct, sep = "_")
      ) +
      theme_classic() +
      facet_wrap(~genotype)
    print(fig)
    ggsave(file.path(fig_dir,paste0(tp,"_",ct,".png",sep='')),fig,width = 12,height = 12)
  }
}

```


# Session information

```{r}
gc()
sessionInfo()
```
